refactor(data): unify snapshot loading logic
